+++ tests/testthat/test-req-perform.R
@@ -103,7 +103,61 @@ test_that("can cache requests with etags", {
   req <- request_test("/etag/:etag", etag = "abc") %>% req_cache(tempfile())
 
   resp1 <- req_perform(req)
-  expect_condition(resp2 <- req_perform(req), class = "httr2_cache_not_modified")
+  expect_condition(
+    expect_condition(resp2 <- req_perform(req), class = "httr2_cache_not_modified"),
+    class = "httr2_cache_save"
+  )
+})
+
+test_that("can cache requests with paths (cache-control)", {
+  req <- request(example_url()) %>%
+    req_url_path("/cache/1") %>%
+    req_cache(tempfile())
+
+  path1 <- tempfile()
+  expect_condition(
+    resp1 <- req %>% req_perform(path = path1),
+    class = "httr2_cache_save"
+  )
+  expect_equal(resp1$body[[1]], path1)
+
+  path2 <- tempfile()
+  expect_condition(
+    resp2 <- req %>% req_perform(path = path2),
+    class = "httr2_cache_cached"
+  )
+  expect_equal(resp2$body[[1]], path2)
+
+  Sys.sleep(1) # wait for cache to expire
+  path3 <- tempfile()
+  expect_condition(
+    resp3 <- req %>% req_perform(path = path3),
+    class = "httr2_cache_save"
+  )
+  expect_equal(resp3$body[[1]], path3)
+})
+
+test_that("can cache requests with paths (if-modified-since)", {
+  req <- request(example_url()) %>%
+    req_url_path("/cache") %>%
+    req_cache(tempfile())
+
+  path1 <- tempfile()
+  expect_condition(
+    resp1 <- req %>% req_perform(path = path1),
+    class = "httr2_cache_save"
+  )
+  expect_equal(resp1$body[[1]], path1)
+
+  path2 <- tempfile()
+  expect_condition(
+    expect_condition(
+      resp2 <- req %>% req_perform(path = path2),
+      class = "httr2_cache_not_modified"
+    ),
+    class = "httr2_cache_save"
+  )
+  expect_equal(resp2$body[[1]], path2)
 })
 
 test_that("can retrieve last request and response", {